pyproject.toml
[tool.poetry]
name = "ska_tmc_simulators"
version = "0.0.7"
description = "This is a package that has simulators used by TMC for testing purposes. It provides simulators for CSP and SDP subsystems for ska telescope."
authors = ["Team HIMALAYA" , "Team SAHYADRI"]
license = "BSD-3-Clause"

packages = [
    { include = "ska_tmc_simulators/cspmastersimulator", from = "src" },
    { include = "ska_tmc_simulators/sdpmastersimulator", from = "src" },
    { include = "ska_tmc_simulators/cspsubarraysimulator", from = "src" },
    { include = "ska_tmc_simulators/sdpsubarraysimulator", from = "src" },
]

[tool.poetry.scripts]
CspMasterSimulatorDS="ska_tmc_simulators.cspmastersimulator.csp_master:main"
CspSubarraySimulatorDS="ska_tmc_simulators.cspsubarraysimulator.csp_subarray:main"
SdpSubarraySimulatorDS="ska_tmc_simulators.sdpsubarraysimulator.sdp_subarray:main"
SdpMasterSimulatorDS="ska_tmc_simulators.sdpmastersimulator.sdp_master:main"

[[tool.poetry.source]]
name = 'ska-nexus'
url = 'https://artefact.skao.int/repository/pypi-internal/simple'

[tool.poetry.dependencies]
python = "^3.8"
pytango = "9.3.3"
marshmallow = "^3.13.0"
fire = "^0.4.0"
ska-ser-logging = "^0.4.1"
ska-tango-base = "0.13.3"
pytz = "2021.3"
ipython = "8.5.0"
tango-simlib = "^0.9.4"
ska_tmc_common = "0.1.54"
pytest-runner = "^6.0.0"
prospector = {extras = ["with_pyroma"], version = "^1.8.4"}
yapf = "^0.32.0"
Sphinx = "^6.1.3"
recommonmark = "^0.7.1"

[tool.poetry.dev-dependencies]
pylint = "^2.12.0"
pytest = "^7.1"
pytest-bdd = "^5.0.0"
pytest-cov = "^2.10.1"
pylint-junit = "^0.3.2"
coverage = "^6.1.1"
pytest-json-report = "^1.4.1"
pytest-forked = "^1.3.0"
pytest-mock = "^3.6.1"
pytest-xdist = "^2.4.0"
pytest-repeat = "^0.9.1"
mock = "^4.0.3"
unittest2 = "1.1.0"
datetime = "4.3"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.10.1"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
